Moved DictListCollection to utils
